ایجاد دکمه جستجو برای نقشه و رفع خطا و ایجاد پیام ها

+++ src/app/api/locations/[id]/route.ts
@@ -13,13 +13,27 @@ export async function PUT(req: Request, { params }: { params: { id: string } }) 
 }
 
 export async function DELETE(req: Request, { params }: { params: { id: string } }) {
-    await prisma.locations.delete({
-        where: { id: Number(params.id) }
+
+
+    const isUsers = await prisma.users.findMany({
+        where: {
+            locationId: Number(params.id),
+        },
     });
-    return NextResponse.json({ message: "Deleted" });
+
+    if (isUsers && isUsers.length !== 0) {
+        return NextResponse.json({ message: "تعدادی از کارمندان به این چایگاه نسبت داده شده اند", done: false });
+    }
+    else {
+        await prisma.locations.delete({
+            where: { id: Number(params.id) }
+        });
+        return NextResponse.json({ message: "عملیات حذف با موفقیت انجام شد", done: true });
+    }
 }
 
-export async function GET(req: Request, { params }: { params: { id: string } }) {
-    const pathInfo = await prisma.paths.findUnique({ where: { id: +params.id } });
+export async function POST(req: Request) {
+    const { id } = await req.json();
+    const pathInfo = await prisma.paths.findUnique({ where: { id: Number(id) } });
     return NextResponse.json(pathInfo);
 }
